// src/sdk/types.ts
export interface Event {
  // 事件id
  id: string
  impl: string
  // 版本
  version: number
  // 平台
  platform: string
  self: {
    // 用户id
    user_id: string
    // 平台
    platform: string
    // 昵称
    nickname: string
    user_displayname: string
  }
  self_id: number
  // 时间错
  time: number
  detail_type: string
  type: string
  sub_type: string
  post_type: string
  message_id: string
  user_id: string
  seq: number
  rand: number
  font: string
  message: Message[]
  raw_message: string
  message_type: string
  sender: {
    user_id: number
    nickname: string
  }
  from_id: number
  to_id: number
  auto_reply: boolean
  friend: {
    user_id: number
    nickname: string
    sex: string
    remark: string
    class_id: number
  }
  cqCode: string
  alt_message: string
}

interface Message {
  type: string
  data: any
}